refactor(l5): move embedding, chunking, embedding-canonical, embedding-targets from src/ into server/embedding (issue 864)
Cluster C3 of the L5 move program. The embedding provider and its helpers are
an outbound adapter, not domain, so they get a new home at server/embedding/.

Moved (old -> new):
- src/chunking.ts -> server/embedding/chunking.ts
- src/chunking.test.ts -> server/embedding/chunking.test.ts
- src/embedding-canonical.ts -> server/embedding/embedding-canonical.ts
- src/embedding-canonical.test.ts -> server/embedding/embedding-canonical.test.ts
- src/embedding-targets.ts -> server/embedding/embedding-targets.ts
- src/embedding-targets.test.ts -> server/embedding/embedding-targets.test.ts

NOT moved: src/embedding.ts stays in src/. It reads nine env keys directly
(EMBEDDING_TIMEOUT_MS, EMBEDDING_DIMENSIONS, EMBEDDING_MODEL,
EMBEDDING_WATCHDOG_FAILURE_THRESHOLD, EMBEDDING_WATCHDOG_COOLDOWN_MS,
EMBEDDING_WATCHDOG_RESTART_SCRIPT, EMBEDDING_BASE_URL, EMBEDDING_API_KEY), and
server/config.ts parses only the last two. EMBEDDING_MODEL and
EMBEDDING_DIMENSIONS are module-scope exported constants consumed as values by
roughly fifty files, so threading them as parameters would edit every one of
those callers, well past the one-import-line allowance for server/main.ts. M4
says stop and report the keys rather than move it.

Each old src/ path is left as an L5 shim re-exporting the new server/ path, so
importers outside the moved set keep their import lines and their pre-existing
lint findings stay out of the staged gate. The shims retire with src/ at L6.

Closure count: 67 before -> 67 after. The shims are themselves src/ files, so
the count holds until L6 removes them; the closure drops by three at that point.

Behavior is unchanged. Two edits were forced by the server/ lint standard:
chunkText's inline break-point search is hoisted into a named findBreakPoint
helper (same arithmetic, same order of preference), and the moved test files
swap non-null assertions for expectDefined and split three oversized describe
blocks into siblings. No test body, name, or assertion changed otherwise.

Verified: bunx tsc --noEmit exit 0; oxlint --deny-warnings clean on every
staged file; bun run test:isolated on the three moved test files plus
server/main.tool-logging.test.ts and server/config*.test.ts -> 154 pass, 0 fail.
